refactored the app and added the episerverdata file to allow others to import the data for the frontend app into their cms instance
