+++ next.config.js
@@ -1,26 +1,25 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-    reactStrictMode: true,
-    eslint: {
-        ignoreDuringBuilds: true,
-    },
-    images: {
-        remotePatterns: [
-            {
-                protocol: 'https',
-                hostname: '*.opti-demo.xyz',
-                pathname: '/globalassets/**',
-            },
-            {
-                protocol: 'https',
-                hostname: '*.welcomesoftware.com',
-            },
-            {
-                protocol: 'https',
-                hostname: 'source.unsplash.com',
-            }
-        ],
-    },
-}
+  reactStrictMode: true,
+  eslint: {
+    ignoreDuringBuilds: true,
+  },
+  images: {
+    remotePatterns: [
+      {
+        protocol: "https",
+        hostname: "*.welcomesoftware.com",
+      },
+      {
+        protocol: "https",
+        hostname: "*.optimizely.com",
+      },
+      {
+        protocol: "https",
+        hostname: "source.unsplash.com",
+      },
+    ],
+  },
+};
 
-module.exports = nextConfig
+module.exports = nextConfig;
