+++ src/app/search/page.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import Header from "@/src/components/Header";
+
+export default function Page() {
+  return (
+    <>
+      <Header height={60} />
+    </>
+  );
+}
