--- src/graphql/pages/BlogList.graphql
@@ -1,27 +1,0 @@
-query BlogList($locale: [Locales] = en){
-    LocationItemPage(locale: $locale, limit: 100, orderBy:{
-        Name : ASC, Saved: DESC
-    }, where: { IsCommonDraft: {eq: true}})
-    {
-        items{
-            Name,
-            RelativePath,
-            StartPublish
-            MainIntro,
-            Image {
-                Url
-            },
-            PageImage {
-                Url
-            },
-            Saved,
-            ContentLink {
-                Id
-                WorkId
-                GuidValue
-                ProviderName
-                Url
-            }
-        }
-    }
-}
