--- src/graphql/pages/BlogPost.graphql
@@ -1,29 +1,0 @@
-query BlogPost($locale: [Locales] = en, $guid: String){
-    LocationItemPage(locale: $locale, where:{
-        ContentLink: {
-            GuidValue: {eq: $guid}
-        }
-    })
-    {
-        items{
-            Name,
-            Url,
-            StartPublish,
-            MainIntro,
-            MainBody,
-            Image {
-                Url
-            },
-            PageImage {
-                Url
-            },
-            ContentLink {
-                Id
-                WorkId
-                GuidValue
-                ProviderName
-                Url
-            }
-        }
-    }
-}
